requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
ultralytics>=8.3.0
opencv-python-headless>=4.10.0.84
aiofiles==23.2.1
aiosqlite==0.20.0
python-dotenv==1.0.1
reportlab==4.2.2
pymodbus==3.6.9
